MAR-213 - added auto generator sitemap
