+++ nuxt.config.js
@@ -1,4 +1,4 @@
-import axios from 'axios'
+import getRoutes, { getSitemapRoutes } from './getRoutes'
 
 require('dotenv').config()
 
@@ -63,66 +63,7 @@ module.exports = {
   },
   components: true,
   generate: {
-    async routes() {
-      const getPosts = async pageUrl => {
-        let posts = []
-        const response = await axios.get(pageUrl)
-        posts = posts.concat(response.data.results)
-
-        if (response.data.next_page) {
-          posts = posts.concat(await getPosts(response.data.next_page))
-        }
-        return posts
-      }
-      const convertToSlug = text => text.toLowerCase()
-        .trim()
-        .replace(/[^\w ]+/g, '')
-        .replace(/ +/g, '-')
-
-      // Getting data from prismic
-      const prismicData = await axios.get(process.env.NODE_PRISMIC_API)
-      const prismicTags = prismicData.data.tags
-
-      // Getting posts of all types from prismic
-      const { ref } = prismicData.data.refs[0]
-      const prismicPosts = await getPosts(`${process.env.NODE_PRISMIC_API}/documents/search?ref=${ref}#format=json`)
-
-      // Creating a list of routes
-      const blogPageRoutes = prismicPosts
-        .filter(post => post.type === 'post')
-        .map(post => `/blog/${post.uid}`)
-
-      const cuPageRoutes = prismicPosts
-        .filter(post => post.type === 'customer_university')
-        .map(post => `/customer-university/${post.uid}`)
-
-      const authorPageRoutes = prismicPosts
-        .filter(post => post.type === 'author')
-        .map(author => `/blog/author/${author.uid}`)
-
-      const tagPageRoutes = prismicTags
-        .map(tag => `/blog/tag/${convertToSlug(tag)}`)
-
-      const routes = [
-        '/',
-        '/services',
-        '/projects',
-        '/careers',
-        '/gdpr',
-        '/nda',
-        '/privacy',
-        '/faq',
-        '/case-studies/namba-food',
-        '/case-studies/sir-john-monash-centre',
-        '/blog',
-        ...blogPageRoutes,
-        ...cuPageRoutes,
-        ...authorPageRoutes,
-        ...tagPageRoutes,
-      ]
-
-      return routes
-    },
+    ...getRoutes(),
     fallback: '404.html',
   },
   css: [
@@ -176,12 +117,13 @@ module.exports = {
   */
   modules: ['@nuxtjs/axios', '@nuxtjs/robots', '@nuxtjs/prismic', '@nuxtjs/sitemap'],
   sitemap: {
-    hostname: 'https://maddevs.io',
+    hostname: process.env.domain,
     gzip: false,
     path: '/sitemap.xml',
-    defaults: {
-      changefreq: 'daily',
-      lastmod: new Date(),
+    exclude: ['/**'],
+    routes: async () => {
+      const routes = await getSitemapRoutes()
+      return routes
     },
   },
   axios: {
